monitor: move unit tests to unit/ folder
