+++ sdk/monitor/opentelemetry-exporter/test/unit/breezeTestUtils.ts
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { BreezeResponse } from "../src/utils/breezeUtils";
+import { BreezeResponse } from "../../src/utils/breezeUtils";
 
 export function successfulBreezeResponse(count: number): BreezeResponse {
   return {
     itemsAccepted: count,
     itemsReceived: count,
-    errors: [],
+    errors: []
   };
 }
 
@@ -18,8 +18,8 @@ export function failedBreezeResponse(count: number, statusCode: number): BreezeR
     errors: new Array(count).fill(0).map((_, index) => ({
       index,
       statusCode,
-      message: "foo",
-    })),
+      message: "foo"
+    }))
   };
 }
 
@@ -33,7 +33,7 @@ export function partialBreezeResponse(statusCodes: number[]) {
       .map((v) => ({
         index: statusCodes.findIndex((s) => v === s),
         statusCode: v,
-        message: "foo",
-      })),
+        message: "foo"
+      }))
   };
 }
