+++ sdk/monitor/opentelemetry-exporter/test/unit/utils/connectionStringParser.test.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import * as Constants from "../../src/Declarations/Constants";
-import { ConnectionStringParser } from "../../src/utils/connectionStringParser";
+import * as Constants from "../../../src/Declarations/Constants";
+import { ConnectionStringParser } from "../../../src/utils/connectionStringParser";
 
 describe("ConnectionStringParser", () => {
   describe("#parse()", () => {
@@ -46,10 +46,12 @@ describe("ConnectionStringParser", () => {
     }) => {
       const result = ConnectionStringParser.parse(options.connectionString);
 
-      if (options.expectedAuthorization)
-        {assert.deepEqual(result.authorization, options.expectedAuthorization);}
-      if (options.expectedInstrumentationKey)
-        {assert.deepEqual(result.instrumentationkey, options.expectedInstrumentationKey);}
+      if (options.expectedAuthorization) {
+        assert.deepEqual(result.authorization, options.expectedAuthorization);
+      }
+      if (options.expectedInstrumentationKey) {
+        assert.deepEqual(result.instrumentationkey, options.expectedInstrumentationKey);
+      }
       assert.deepEqual(result.ingestionendpoint, options.expectedBreezeEndpoint);
       assert.deepEqual(result.liveendpoint, options.expectedLiveMetricsEndpoint);
     };
@@ -60,7 +62,7 @@ describe("ConnectionStringParser", () => {
         expectedAuthorization: undefined,
         expectedInstrumentationKey: "00000000-0000-0000-0000-000000000000",
         expectedBreezeEndpoint: Constants.DEFAULT_BREEZE_ENDPOINT,
-        expectedLiveMetricsEndpoint: Constants.DEFAULT_LIVEMETRICS_ENDPOINT,
+        expectedLiveMetricsEndpoint: Constants.DEFAULT_LIVEMETRICS_ENDPOINT
       });
     });
 
@@ -69,7 +71,7 @@ describe("ConnectionStringParser", () => {
         connectionString:
           "InstrumentationKey=00000000-0000-0000-0000-000000000000;EndpointSuffix=ai.contoso.com",
         expectedBreezeEndpoint: "https://dc.ai.contoso.com",
-        expectedLiveMetricsEndpoint: "https://live.ai.contoso.com",
+        expectedLiveMetricsEndpoint: "https://live.ai.contoso.com"
       });
     });
 
@@ -78,7 +80,7 @@ describe("ConnectionStringParser", () => {
         connectionString:
           "InstrumentationKey=00000000-0000-0000-0000-000000000000;EndpointSuffix=ai.contoso.com;LiveEndpoint=https://custom.live.contoso.com:444",
         expectedBreezeEndpoint: "https://dc.ai.contoso.com",
-        expectedLiveMetricsEndpoint: "https://custom.live.contoso.com:444",
+        expectedLiveMetricsEndpoint: "https://custom.live.contoso.com:444"
       });
     });
 
@@ -87,7 +89,7 @@ describe("ConnectionStringParser", () => {
         connectionString:
           "InstrumentationKey=00000000-0000-0000-0000-000000000000;EndpointSuffix=ai.contoso.com;Location=westus2",
         expectedBreezeEndpoint: "https://westus2.dc.ai.contoso.com",
-        expectedLiveMetricsEndpoint: "https://westus2.live.ai.contoso.com",
+        expectedLiveMetricsEndpoint: "https://westus2.live.ai.contoso.com"
       });
     });
 
@@ -96,7 +98,7 @@ describe("ConnectionStringParser", () => {
         connectionString:
           "InstrumentationKey=00000000-0000-0000-0000-000000000000;EndpointSuffix=ai.contoso.com;Location=westus2;LiveEndpoint=https://custom.contoso.com:444",
         expectedBreezeEndpoint: "https://westus2.dc.ai.contoso.com",
-        expectedLiveMetricsEndpoint: "https://custom.contoso.com:444",
+        expectedLiveMetricsEndpoint: "https://custom.contoso.com:444"
       });
     });
 
@@ -105,7 +107,7 @@ describe("ConnectionStringParser", () => {
         connectionString:
           "InstrumentationKey=00000000-0000-0000-0000-000000000000;LiveEndpoint=http://custom.live.endpoint.com:444",
         expectedBreezeEndpoint: Constants.DEFAULT_BREEZE_ENDPOINT,
-        expectedLiveMetricsEndpoint: "http://custom.live.endpoint.com:444",
+        expectedLiveMetricsEndpoint: "http://custom.live.endpoint.com:444"
       });
     });
   });
